requirements.txt
fastmcp>=2.0.0
requests
google-auth
google-auth-oauthlib
